Add token and enterprise variables to usage.yaml workflow
This commit adds the `token` and `enterprise` variables to the `usage.yaml` workflow. The `token` variable is set to the value of the `secrets.PAT` secret, and the `enterprise` variable is set to the value of the `vars.enterprise` variable. This allows the workflow to access the necessary credentials for authentication.