Add startGroup and endGroup functions to @actions/core
This commit adds the startGroup and endGroup functions to the @actions/core module. It also imports the info function from the console module. The startGroup function is used to start a new group in the GitHub Actions workflow, while the endGroup function is used to end a group. The info function is used to log information to the console.